# docs/openapi/argos.v1.yaml
openapi: 3.0.3
info:
  title: ARGOS API
  version: 1.0.0
  description: Contrato inicial da API versionada do ARGOS.
servers:
  - url: http://localhost:3333/api/v1
security:
  - bearerAuth: []
paths:
...
  /items:
    get:
      tags: [Items]
      summary: Lista itens do usuário autenticado.
      responses:
        '200':
          description: Itens privados do usuário.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PrivateItem'
    post:
      tags: [Items]
      summary: Publica item aprovado com data do ocorrido informada ou data atual.
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          description: Chave por usuário de 16 a 100 caracteres. Repetir o mesmo payload devolve o mesmo ID; mudar o payload retorna 409. Guarde a chave ao receber timeout.
          schema: { type: string, pattern: '^[a-zA-Z0-9_-]{16,100}$' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateItemRequest'
      responses:
        '201':
          description: Item criado; resposta contém id e message.
        '200':
          description: Operação repetida; resposta contém id original, message e replayed verdadeiro.
        '409':
          $ref: '#/components/responses/Conflict'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /admin/items:
    get:
      tags: [Admin]
      summary: Lista itens para moderação global.
      parameters:
        - name: approvalStatus
          in: query
          schema:
            $ref: '#/components/schemas/ApprovalStatus'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ItemStatus'
        - name: q
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Itens para administração.
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/items/{id}/status:
    patch:
      tags: [Admin]
      summary: Atualiza aprovação ou status operacional.
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                approvalStatus:
                  $ref: '#/components/schemas/ApprovalStatus'
                status:
                  $ref: '#/components/schemas/ItemStatus'
                moderationNote:
                  type: string
                  maxLength: 500
      responses:
        '200':
          description: Status atualizado.
...
  /admin/users/{id}:
    patch:
      tags: [Admin]
      summary: Atualiza role, status ou score anti-spam.
      description: Bloquear ou tornar pendente revoga imediatamente todos os refresh tokens do usuário.
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role:
                  type: string
                  enum: [citizen, space_manager, org_admin, support, admin]
                status:
                  type: string
                  enum: [pending, active, blocked]
                spamScore:
                  type: integer
                  minimum: 0
                  maximum: 10
      responses:
        '200':
          description: Usuário atualizado.
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          description: Dados inválidos ou tentativa do administrador de remover o próprio acesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /admin/access-requests/{id}:
    patch:
      tags: [Admin]
      summary: Aprova ou rejeita solicitação de acesso.
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Solicitação revisada.
...
  /admin/mail-test:
    post:
      tags: [Admin]
      summary: Confere a conexão SMTP e envia um e-mail de teste ao administrador.
      responses:
        '200':
          description: Resultado (configured true/false).
        '502':
          description: Servidor SMTP recusou a conexão.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: argos_refresh
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    ItemType:
      type: string
      enum: [lost, found]
    ItemStatus:
      type: string
      enum: [lost, found, claimed, returned]
    ApprovalStatus:
      type: string
      enum: [pending, approved, rejected]
    PublicItem:
      type: object
      required: [id, type, title, description, category, location, event_date, status, approval_status, created_at]
      properties:
        id:
          type: integer
        type:
          $ref: '#/components/schemas/ItemType'
        title:
          type: string
        description:
          type: string
        category:
          type: string
        location:
          type: string
        approximate_place:
          type: string
          nullable: true
        event_date:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/ItemStatus'
        approval_status:
          $ref: '#/components/schemas/ApprovalStatus'
        image_url:
          type: string
          nullable: true
        owner_nickname:
          type: string
        owner_avatar_url:
          type: string
          nullable: true
        created_at:
          type: string
        comments_count:
          type: integer
        latest_comments:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
    PrivateItem:
      allOf:
        - $ref: '#/components/schemas/PublicItem'
        - type: object
          properties:
            owner_id:
              type: integer
            owner_name:
              type: string
            campus_block:
              type: string
              nullable: true
            contact_preference:
              type: string
              enum: [in_app, email]
            moderation_note:
              type: string
              nullable: true
            updated_at:
              type: string
    CreateItemRequest:
      type: object
      required: [type, title, description, category, location]
      properties:
        type:
          $ref: '#/components/schemas/ItemType'
        title:
          type: string
          minLength: 3
          maxLength: 120
        description:
          type: string
          minLength: 10
          maxLength: 2000
        category:
          type: string
        location:
          type: string
        campusBlock:
          type: string
        approximatePlace:
          type: string
        eventDate:
          type: string
          format: date
          description: Campo legado opcional; se enviado, não pode ser futuro. A data gravada é definida pelo servidor.
        imageUrl:
          type: string
        contactPreference:
          type: string
          enum: [in_app, email]
...
    Comment:
      type: object
      required: [id, item_id, user_id, author_name, body, created_at]
      properties:
        id:
          type: integer
        item_id:
          type: integer
        user_id:
          type: integer
        author_name:
          type: string
        author_nickname:
          type: string
        author_avatar_url:
          type: string
          nullable: true
        body:
          type: string
        created_at:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        details:
          type: object
        requestId:
          type: string
...
  responses:
    Unauthorized:
      description: Não autenticado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Sem permissão.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Recurso não encontrado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflito.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ValidationError:
      description: Dados inválidos.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'